Build fixes for 64-bit systems, partial fixes from James Henstridge
2005-06-11  Bastien Nocera  <hadess@hadess.net>

	* player/rb-recorder-gst.c: (rb_recorder_burn):
	* sources/rb-playlist-source-recorder.c:
	(rb_playlist_source_recorder_start):
	* widgets/rb-header.c: (info_url): Build fixes for 64-bit systems,
	partial fixes from James Henstridge <james@jamesh.id.au>
	(Closes: #168727)

+++ widgets/rb-header.c
@@ -474,7 +474,8 @@ info_url (guint opt, const char *data)
 	} else if (g_str_has_prefix (data, "www.")) {
 		char *end = strchr (data, ' ');
 		if (end != NULL)
-			ret = g_strdup_printf ("http://%*s", end-data, data);
+			ret = g_strdup_printf ("http://%*s",
+					       (int) (end-data), data);
 		else
 			ret = g_strdup_printf ("http://%s", data);
 	} else {
